requirements.txt
yt-dlp>=2024.1.0
youtube-transcript-api>=0.6.2
anthropic>=0.25.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
